Function out mapping for market-period
